fix: 修复构建失败使测试无法执行，移除 continue-on-error 掩盖
根因：
- Sandbox.swift:29,44: Int→UInt 转换遗漏（aether_sandbox_new/load 的 uintptr_t 参数）
- RateLimiter.swift:60,72: Int→TimeInterval 类型不匹配（rateLimited 参数是 Double）
- CI 构建步骤 continue-on-error: true 掩盖了构建失败，导致测试步骤在
  未构建成功的 app 上运行，所有测试实际未执行（Missing bundle ID）

修复：
- Sandbox.swift: UInt(maxMemoryBytes), UInt(count)
- RateLimiter.swift: 移除 Int() 包裹，retryAfter.rounded(.up) 已是 Double
- CI: 移除所有 5 个 Build for Testing 步骤的 continue-on-error: true 和 set +e
  构建失败将正确阻止后续测试步骤，不再虚假通过

+++ Packages/AetherCore/Sources/AetherRust/Sandbox.swift
@@ -26,7 +26,7 @@ public final class AetherRustSandbox: @unchecked Sendable {
     ///   - maxFuel: CPU 指令限额（30 秒 ≈ 30_000_000_000）
     ///   - maxMemoryBytes: 线性内存上限（字节，50 MB = 52_428_800）
     public init?(maxFuel: UInt64 = 30_000_000_000, maxMemoryBytes: Int = 50 * 1024 * 1024) {
-        guard let h = aether_sandbox_new(maxFuel, maxMemoryBytes) else { return nil }
+        guard let h = aether_sandbox_new(maxFuel, UInt(maxMemoryBytes)) else { return nil }
         self.handle = h
     }
 
@@ -41,7 +41,7 @@ public final class AetherRustSandbox: @unchecked Sendable {
         guard count > 0 else { return nil }
         let module = wasm.withUnsafeBytes { (ptr: UnsafeRawBufferPointer) -> OpaquePointer? in
             guard let base = ptr.baseAddress else { return nil }
-            return aether_sandbox_load(handle, base.assumingMemoryBound(to: UInt8.self), count)
+            return aether_sandbox_load(handle, base.assumingMemoryBound(to: UInt8.self), UInt(count))
         }
         guard let m = module else { return nil }
         return AetherRustSandboxModule(handle: m)
